+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - introducing `withLock` wrapper
+- introducing `join` command
 - improved error messages by using command name prefix (fix #88)
 - multiline pipelines (fix #83)